+++ example/src/components/demo.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { useForm } from 'slimeform'
 
-const { form, status, reset, onSubmit, clearErrors } = useForm({
+const { form, status, reset, submitter, clearErrors, isError, rule } = useForm({
   // Initial form value
   form: () => ({
     age: '',
@@ -18,9 +18,22 @@ const { form, status, reset, onSubmit, clearErrors } = useForm({
   defaultMessage: '\u00A0',
 })
 
-function mySubmit() {
+const { submit, submitting } = submitter(async ({ form }) => {
+  await new Promise(resolve => setTimeout(resolve, 1000))
   // eslint-disable-next-line no-alert
   alert(`Age: ${form.age}`)
+})
+
+const preInput = ref('')
+const preInputStatus = ref({ valid: true, message: '\u00A0' })
+function validatePreInput() {
+  const { valid, message } = rule.age.validate(preInput.value, { fullResult: true })
+  preInputStatus.value = { valid, message }
+
+  if (valid) {
+    form.age = preInput.value
+    preInput.value = ''
+  }
 }
 </script>
 
@@ -29,8 +42,9 @@ function mySubmit() {
     Please enter your age
   </h3>
 
-  <form @submit.prevent="onSubmit(mySubmit)">
-    <label>
+  <form space="y-2" @submit.prevent="submit">
+    <!-- Input Age -->
+    <label class="block">
       <input
         v-model="form.age"
         placeholder="What's your age?"
@@ -46,9 +60,42 @@ function mySubmit() {
       >
       <p class="text-red !mb-0 !mt-1 text-sm">{{ status.age.message }}</p>
     </label>
-    <div space-x-3 mt-1>
-      <button text-sm btn type="submit">
-        Submit
+
+    <div flex="~" justify="center" text="xl">
+      <div class="i-carbon:arrow-up" />
+    </div>
+
+    <!-- Test Validate -->
+    <label class="block">
+      <div space="x-2" flex="~" justify="center">
+        <input
+          v-model="preInput"
+          placeholder="Pre input"
+          type="text"
+          autocomplete="false"
+          p="x-4 y-1"
+          w="165px"
+          text="center sm"
+          bg="transparent"
+          border="~ rounded gray-200 dark:gray-700"
+          outline="none active:none"
+          :class="!preInputStatus.valid && '!border-red'"
+        >
+        <button text-sm btn type="button" @click="validatePreInput()">
+          Validate
+        </button>
+      </div>
+      <p class="text-red !mb-0 !mt-1 text-sm">{{ preInputStatus.message }}</p>
+    </label>
+
+    <!-- Buttons -->
+    <div space-x-3>
+      <button text-sm btn type="submit" :disabled="isError || submitting">
+        {{
+          submitting
+            ? 'Submitting...'
+            : 'Submit'
+        }}
       </button>
 
       <button text-sm btn type="button" @click="clearErrors">