+++ src/index.test.ts
@@ -423,6 +423,30 @@ describe('useForm', () => {
     expect(wrapper.isError).toBe(true)
     expect(wrapper.status.age.message).toBe('expect numbers')
   })
+
+  it('returns the rule object', async () => {
+    const wrapper = useSetup(() => {
+      const { form, rule } = useForm({
+        form: () => ({
+          name: '',
+          age: '',
+        }),
+        rule: {
+          age: val => !isNaN(+val) || 'expect numbers',
+          name: v => v.length < 3 || 'to many characters',
+        },
+        defaultMessage: '\u00A0',
+      })
+      return { form, rule }
+    })
+
+    expect(wrapper.rule.age.validate(20)).toBe(true)
+    expect(wrapper.rule.age.validate(20, { fullResult: true })).toStrictEqual({ valid: true, message: '\u00A0' })
+    expect(wrapper.rule.age.validate(20, { fullResult: false })).toEqual(true)
+    expect(wrapper.rule.age.validate('abc')).toBe(false)
+    expect(wrapper.rule.age.validate('abc', { fullResult: true })).toStrictEqual({ valid: false, message: 'expect numbers' })
+    expect(wrapper.rule.age.validate('abc', { fullResult: false })).toEqual(false)
+  })
 })
 
 describe('object type field', () => {